Refactor domain models and update migrations
- Added `GetAllWithSpec` and `GetByIdWithSpec` methods to `IGenericEntity` and `GenericEntityRepo` for specification-based queries.
- Changed `Price` property type from `decimal` to `int` in `Product` and updated related migrations.
- Removed outdated `20251029224443_CreateDatabase` migration files.
- Added new `20251111125334_CreateDabase` migration to reflect schema changes.
- Updated `StoreHubDbContextModelSnapshot` to include lazy loading and proxy annotations.
- Updated `appsettings.json` connection string to use a local SQL Server instance.
- Applied minor formatting fixes across the codebase.

+++ Core/Domain/Models/Product.cs
@@ -7,12 +7,13 @@ using System.Threading.Tasks;
 
 namespace Domain.Models
 {
-    public class Product :BaseEntity<int>
+    public class Product : BaseEntity<int>
     {
         public string Name { get; set; }
         public string Description { get; set; }
         public string PictureUrl { get; set; }
-        public decimal Price { get; set; }
+
+        public int Price { get; set; }
         [ForeignKey(nameof(ProductBrands))]
         public int BrandId { get; set; }
         public virtual ProductBrand ProductBrands { get; set; }
